# src/main/java/com/IlyaJukov/SocialMedia/models/request/SignInRequest.java
package com.IlyaJukov.SocialMedia.models.request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data @Builder
@NoArgsConstructor @AllArgsConstructor
public class SignInRequest {
    private String login;
    private String password;
}