# src/main/java/com/IlyaJukov/SocialMedia/models/request/SignUpRequest.java
package com.IlyaJukov.SocialMedia.models.request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data @Builder
@NoArgsConstructor @AllArgsConstructor
public class SignUpRequest {
    private String name;
    private String login;
    private String password;
}